+++ src/translation/src/Contracts/Translator.php
@@ -17,35 +17,35 @@ interface Translator
     /**
      * Get the translation for a given key.
      *
-     * @param string $key
-     * @param array $replace
+     * @param string      $key
+     * @param array       $replace
      * @param null|string $locale
      * @return mixed
      */
-    public function trans($key, array $replace = [], $locale = null);
+    public function trans(string $key, array $replace = [], $locale = null);
 
     /**
      * Get a translation according to an integer value.
      *
-     * @param string $key
+     * @param string               $key
      * @param array|\Countable|int $number
-     * @param array $replace
-     * @param null|string $locale
+     * @param array                $replace
+     * @param null|string          $locale
      * @return string
      */
-    public function transChoice($key, $number, array $replace = [], $locale = null);
+    public function transChoice(string $key, $number, array $replace = [], $locale = null): string;
 
     /**
      * Get the default locale being used.
      *
      * @return string
      */
-    public function getLocale();
+    public function getLocale(): string;
 
     /**
      * Set the default locale.
      *
      * @param string $locale
      */
-    public function setLocale($locale);
+    public function setLocale(string $locale);
 }